+++ src/styles/components/checkbox-style.js
@@ -2,29 +2,33 @@ const csjs = require('csjs-inject');
 
 import { theme } from '../../theme.js';
 
+let borderSize = '4px'
+
 module.exports = csjs`
 
 input[type=checkbox].guify-checkbox {
+    display: none;
+    opacity: 0;
     appearance: none;
     -moz-appearance: none;
     -webkit-appearance: none;
     margin: 0;
+    border-radius: 0;
     cursor: pointer;
 }
 
-input[type=checkbox] + label {
+input[type=checkbox].guify-checkbox + label {
     margin: 0;
 }
 
 input[type=checkbox].guify-checkbox + label:before {
     content: "";
     display: inline-block;
-    width: 18px;
-    height: 18px;
-    padding: 0px;
+    width: ${theme.sizing.componentHeight};
+    height: ${theme.sizing.componentHeight};
+    padding: 0;
+    margin: 0;
     vertical-align: middle;
-    margin-right: 8px;
-    margin-top: 1px;
     background-color: ${theme.colors.componentBackground};
     border-radius: 0px;
     cursor: pointer;
@@ -36,26 +40,26 @@ input[type=checkbox].guify-checkbox + label:before {
 
 /* Hover style */
 input[type=checkbox].guify-checkbox:hover + label:before {
-    width: 10px;
-    height: 10px;
+    width: calc(${theme.sizing.componentHeight} - (${borderSize} * 2));
+    height: calc(${theme.sizing.componentHeight} - (${borderSize} * 2));
     background-color: ${theme.colors.componentBackgroundHover};
     border: solid 4px ${theme.colors.componentBackground};
 }
 
 /* Checked style */
 input[type=checkbox]:checked.guify-checkbox + label:before {
-    width: 10px;
-    height: 10px;
+    width: calc(${theme.sizing.componentHeight} - (${borderSize} * 2));
+    height: calc(${theme.sizing.componentHeight} - (${borderSize} * 2));
     background-color: ${theme.colors.componentForeground};
-    border: solid 4px ${theme.colors.componentBackground};
+    border: solid ${borderSize} ${theme.colors.componentBackground};
 }
 
 /* Focused and checked */
 input[type=checkbox]:checked.guify-checkbox:focus + label:before {
-    width: 10px;
-    height: 10px;
+    width: calc(${theme.sizing.componentHeight} - (${borderSize} * 2));
+    height: calc(${theme.sizing.componentHeight} - (${borderSize} * 2));
     background-color: ${theme.colors.componentForeground};
-    border: solid 4px ${theme.colors.componentBackgroundHover};
+    border: solid ${borderSize} ${theme.colors.componentBackgroundHover};
 }
 
 /* Focus and unchecked */