+++ src/styles/container-style.js
@@ -5,10 +5,11 @@ import { theme } from '../theme.js';
 module.exports = csjs`
 
 .guify-container {
-    position: absolute;
+    position: relative;
     left: 0;
     width: 100%;
     font-size: 11px;
+    z-index: 9999;
 }
 
 `;