+++ src/toast-area.js
@@ -24,55 +24,30 @@ export class ToastArea {
      * Makes a message that appears under the menu bar. Transitions out
      * over `transitionMS` milliseconds after `stayMS` milliseconds have passed.
      */
-    CreateToast(message, stayMS = 5000, transitionMS = 100) {
+    CreateToast(message, stayMS = 5000, transitionMS = 0) {
         console.log('[Toast] ' + message);
 
         let toast = this.element.appendChild(document.createElement('div'));
         toast.classList.add(this.styles['guify-toast-notification']);
         toast.setAttribute('aria-live', 'polite');
 
-        css(toast, {
-            'box-sizing': 'border-box',
-            //'background-color': theme.background2,
-            'color': theme.colors.text1,
-            'position': 'relative',
-            'width': '100%',
-            //'height': '20px',
-            'padding': '8px',
-            'padding-left': '20px',
-            'padding-right': '20px',
-            'text-align': 'center',
-            'font-family': `Hack', monospace`,
-            'font-size': '11px',
+        toast.innerHTML = message;
 
+        css(toast, {
             // Animation stuff
             // '-webkit-transition': 'opacity ' + transitionMS + 'ms linear',
             // 'transition': 'opacity ' + transitionMS + 'ms linear',
         });
 
-        toast.innerHTML = message;
-
         // Make close button in toast
         let closeButton = toast.appendChild(document.createElement('button'));
         closeButton.innerHTML = '&#10006;'
-        css(closeButton, {
-            background: 'rgba(0, 0, 0, 0)',
-            'color': theme.colors.text1,
-            position: 'absolute',
-            textAlign: 'center',
-            'margin-top': 'auto',
-            'margin-bottom': 'auto',
-            border: 'none',
-            cursor: 'pointer',
-            fontFamily: 'inherit',
-            top: '0',
-            bottom: '0',
-            right: '8px',
-        })
+        closeButton.classList.add(this.styles['guify-toast-close-button']);
 
         let timeout;
 
         let TransitionOut = () => {
+            toast.blur();
             css(toast, {
                 //'transform-style': 'flat',
                 //'transform-style': 'preserve-3d',
@@ -84,8 +59,8 @@ export class ToastArea {
                 // 'transform:': 'translate3d(0, -100%, 0)',
 
                 // Fade out
-                '-webkit-transition': '-webkit-opacity ' + transitionMS + 'ms linear',
-                'transition': 'opacity ' + transitionMS + 'ms linear',
+                //'-webkit-transition': '-webkit-opacity ' + transitionMS + 'ms linear',
+                //'transition': 'opacity ' + transitionMS + 'ms linear',
                 'opacity': '0',
             });
             clearTimeout(timeout);